Made most drops rarer, in preparation for next time.
This config will be built into the *next* release, probably for a 2017 run.

On a long run of this plugin you end up with an absurd amount of stuff,
particularly Cranberry Sauce, which doesn't stack. The drop rate of Wishbones
could also be reduced, but was kept the same for the sake of comparability.

For planning purposes, estimate up to 1000 mobs killed per player per day, on
average, sustained over a long event (players can kill more per day, but the
shine wears off after a few days).